requirements.txt
Django==1.7
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.3
Pygments==1.6
Sphinx==1.2.3
django-appconf==0.6
django-ckeditor-updated==4.4.4
django-classy-tags==0.5.1
django-compressor==1.4
django-debug-toolbar==1.2.1
django-debug-toolbar-template-timings==0.6.4
django-extensions==1.4.0
django-grappelli==2.6.1
django-hvad==0.5.0
docutils==0.12
easy-thumbnails==2.1
html5lib==0.999
lxml==3.4.0
memcache-toolbar==0.5.5
python-dateutil==2.2
python3-memcached==1.51
six==1.8.0
sqlparse==0.1.12
